refactor: 오버레이 셋팅 및 이동하는 함수 helper function 분리 #46
